Moving pkg/kubectl/util/storage to staging
