+++ pkg/kubectl/describe/versioned/BUILD
@@ -13,7 +13,6 @@ go_library(
         "//pkg/kubectl/util/qos:go_default_library",
         "//pkg/kubectl/util/resource:go_default_library",
         "//pkg/kubectl/util/slice:go_default_library",
-        "//pkg/kubectl/util/storage:go_default_library",
         "//staging/src/k8s.io/api/apps/v1:go_default_library",
         "//staging/src/k8s.io/api/autoscaling/v1:go_default_library",
         "//staging/src/k8s.io/api/autoscaling/v2beta2:go_default_library",
@@ -48,6 +47,7 @@ go_library(
         "//staging/src/k8s.io/kubectl/pkg/util/deployment:go_default_library",
         "//staging/src/k8s.io/kubectl/pkg/util/event:go_default_library",
         "//staging/src/k8s.io/kubectl/pkg/util/rbac:go_default_library",
+        "//staging/src/k8s.io/kubectl/pkg/util/storage:go_default_library",
         "//vendor/github.com/fatih/camelcase:go_default_library",
         "//vendor/k8s.io/klog:go_default_library",
     ],
