+++ pkg/kubectl/describe/versioned/describe.go
@@ -68,6 +68,7 @@ import (
 	deploymentutil "k8s.io/kubectl/pkg/util/deployment"
 	"k8s.io/kubectl/pkg/util/event"
 	"k8s.io/kubectl/pkg/util/rbac"
+	storageutil "k8s.io/kubectl/pkg/util/storage"
 	"k8s.io/kubernetes/pkg/kubectl/describe"
 	"k8s.io/kubernetes/pkg/kubectl/scheme"
 	"k8s.io/kubernetes/pkg/kubectl/util/certificate"
@@ -75,7 +76,6 @@ import (
 	"k8s.io/kubernetes/pkg/kubectl/util/qos"
 	resourcehelper "k8s.io/kubernetes/pkg/kubectl/util/resource"
 	"k8s.io/kubernetes/pkg/kubectl/util/slice"
-	storageutil "k8s.io/kubernetes/pkg/kubectl/util/storage"
 )
 
 // Each level has 2 spaces for PrefixWriter
