+++ staging/src/k8s.io/kubectl/pkg/util/storage/BUILD
@@ -3,7 +3,8 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["storage.go"],
-    importpath = "k8s.io/kubernetes/pkg/kubectl/util/storage",
+    importmap = "k8s.io/kubernetes/vendor/k8s.io/kubectl/pkg/util/storage",
+    importpath = "k8s.io/kubectl/pkg/util/storage",
     visibility = ["//visibility:public"],
     deps = [
         "//staging/src/k8s.io/api/core/v1:go_default_library",
